feat: made write rate configurable (#250)
This makes the write rate configurable by adding writeRate to the limit configuration. Write rate allows to rate-limit the underlying socket flushes and greatly (15-20x from our benchmarks) increases the througput of emitter broker. The value is set to a reasonnable 60 write/s per connection. Keep in mind this is not limiting the QpS but merely limiting the amount of syscalls going out to the OS.