Added last will & refactored service (#321)
This PR adds an implementation of "Last Will & Testament" of MQTT to emitter. This handles 2 scenarios:

 1. When client is terminated but the broker is alive.
 2. When the broker is terminated.

In both cases, last will will be sent to the cluster and optionally message will be retained. This uses consistent hashing to determine which peer will send out the last will in case of the broker termination.

Moreover, I have done a massive refactoring of the service (long overdue). In this refactoring, all of the handlers and services have been made simpler with clear dependencies (see service/ folder). In addition, all tests for the handlers have been rewritten and they have 100% code coverage.