Limit the size of a response
From the begining, mqtt.go defines a maximum size for a response.
However if the limit was met, it resulted in an error and no message was
sent. Now, the Query itself limits the size of a frame and messages are
sent to the client, but only those whose total size <= to the total size
of allowed in a response.

Implications on the current API: for a subscribe request with a Last=X
parameter, the client will receive the last messages whose total size is <= to
the max size of a response, with a maximum of X messages. This behavior is an
improvement over just getting an error.