Adds the new history API
The history api allows to retrieve any number of messages stored in a channel.
As the number of messages that can be returned in a single response is limited
by the size of a response (as per #413), a system of pagination is required.
This system is implemented with the introduction of the StartFromID query
parameter. An client that already received a response with a certain number of
messages can send a new request to the API with the StartFromID set to the ID
of the oldest message it received (first in the array). The storage query will
then begin to retrieve messages starting at the ID preceding the ID provided in
the query, to generate the next page of messages.