[IMP] Adds a Status (200) field to Me and Presence response for consistency
NOTE that we also send a Response with 200 in the case of a response to
a Presence request with the status flag set to false. In this case, we
don't gather the presence statuses on the brokers, but we need to send
an empty response instead of just nil. This will allow for an RPC call
on Presence. See the SDKs (Go in particular).